% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_sampler.R
\name{slice_sampler}
\alias{slice_sampler}
\title{Markov Chain Monte Carlo with Slice Sampling}
\usage{
slice_sampler(
  f,
  x_init,
  w,
  n_samples = 1000,
  max_extensions = NULL,
  procedure_univ = NULL,
  f_log_form = TRUE,
  param_log = NULL
)
}
\arguments{
\item{f}{Density to sample from.}

\item{x_init}{Numerical vector containing initial state of chain.}

\item{w}{Scale estimate. For more information see details section.}

\item{n_samples}{Number of samples to be drawn.}

\item{max_extensions}{Maximal interval extensions in interval search for univariate
slice sampling. Does not affect multivariate slice sampling.}

\item{procedure_univ}{Interval search procedure for univariate slice sampling.
Can be either `stepping_out` (default) or `doubling`.
Does not affect multivariate slice sampling.}

\item{f_log_form}{\code{TRUE} if \code{f} is already in log form (e.g.
if \code{f} is log_likelihood). Default is \code{TRUE}.}

\item{param_log}{\code{TRUE} if temporary parameters such as horizontal value \code{y}
should be saved. Those can be used for visualizations.}
}
\value{
vector (in univariate case) or matrix (in multivariate case) containing
the drawn samples from \code{f}.
}
\description{
Samples from a function using Slice Sampling introduced by Neal (2003).
Depending on \code{x_curr} either univariate or multivariate slice sampling
is performed. In the univariate case, the user can choose between the
stepping out and the doubling procedure for interval search.
}
\details{
In the univariate case param \code{w} determines the interval width for interval search.
In the multivariate case its components contain the side lengths of the hyperrectangle
for all variables in \code{x_curr}.
}
\references{
\insertRef{neal2003slice}{bisSliceSampling}
}
